fix: update memory management and function signatures

+++ core/exec.c
@@ -33,9 +33,9 @@ static char** store_argv(char* path, char** args) {
 static void free_argv(char** args) {
     u32 i;
     for(i=0; args[i] != NULL; i++) {
-        free_mem(args[i]);
+        free_mem((u32)args[i]);
     }
-    free_mem(args);
+    free_mem((u32)args);
 }
 
 u32 ustack_push(u32* esp, char* buf, u32 len) {
